fix lint and tests
