+++ web/src/engine/platform/RemoteBrowserWindow.ts
@@ -1,10 +1,10 @@
+import type { IObservable } from '../Observable';
+import type { ScriptInjection } from './FetchProviderCommon';
 import { Runtime } from './PlatformInfo';
 import { PlatformInstanceActivator } from './PlatformInstanceActivator';
 import NodeWebkitRemoteBrowserWindow from './nw/RemoteBrowserWindow';
 import ElectronRemoteBrowserWindow from './electron/RemoteBrowserWindow';
 import GetIPC from './InterProcessCommunication';
-import type { IObservable } from '../Observable';
-import type { ScriptInjection } from './FetchProvider';
 
 export interface IRemoteBrowserWindow {
     get DOMReady(): IObservable<void, IRemoteBrowserWindow>;
